+++ R/osat.R
@@ -58,7 +58,9 @@ osat_score <- function(bc, batch_vars, feature_vars, expected_dt = NULL, quiet =
   special_col_names <- c(".n_batch", ".batch_freq", ".n_samples")
   special_col_names_str <- stringr::str_c(special_col_names, collapse = ", ")
   assertthat::assert_that(length(intersect(special_col_names, colnames(df))) == 0,
-    msg = glue::glue("special names ({special_col_names_str}) cannot be used as column names")
+    msg = stringr::str_glue(
+      "special names ({special_col_names_str}) cannot be used as column names"
+    )
   )
   if (is.null(expected_dt)) {
     batch_df <- df[, .(.n_batch = .N), by = batch_vars]
@@ -86,12 +88,14 @@ osat_score <- function(bc, batch_vars, feature_vars, expected_dt = NULL, quiet =
     data.table::setkeyv(expected_dt, c(batch_vars, feature_vars))
   } else {
     assertthat::assert_that(is.data.frame(expected_dt) && nrow(expected_dt) > 0)
-    expected_colnames <- c(feature_vars, batch_vars) %>%
-      c(".n_expected") %>%
+    expected_colnames <- c(feature_vars, batch_vars) |>
+      c(".n_expected") |>
       sort()
     expected_colnames_str <- stringr::str_c(expected_colnames, collapse = ", ")
     assertthat::assert_that(all(sort(colnames(expected_dt)) == expected_colnames),
-      msg = glue::glue("expecting column names in expected_dt: {expected_colnames_str}")
+      msg = stringr::str_glue(
+        "expecting column names in expected_dt: {expected_colnames_str}"
+      )
     )
   }
   sample_count_df <- na.omit(df[, .N, by = c(feature_vars, batch_vars)])