+++ R/permute_subgroups.R
@@ -35,7 +35,7 @@ form_homogeneous_subgroups <- function(batch_container, allocate_var, keep_toget
   # The allocation variable must have factor levels in a given order!
   # This is important later on for sample swapping as it has to be known which factor level corresponds to 'group 1' etc
   if (!inherits(allocate_fac, "factor")) {
-    allocate_fac <- forcats::as_factor(allocate_fac)
+    allocate_fac <- factor(allocate_fac, levels = unique(allocate_fac))
   }
 
   assertthat::assert_that(length(keep_together_vars) > 0, msg = "Function can only help if vector of 'keep_together_vars' is specified.")
@@ -104,12 +104,12 @@ form_homogeneous_subgroups <- function(batch_container, allocate_var, keep_toget
   }
 
   # Group sample list by relevant variables
-  grouped_samples <- dplyr::group_by(samples, dplyr::across(tidyselect::all_of(use_vars)))
+  grouped_samples <- dplyr::group_by(samples, dplyr::across(dplyr::all_of(use_vars)))
 
   # Determine sizes of the subgroups and store in list; name elements by levels of the involved grouping variables
   subgroup_sizes <- purrr::map(dplyr::group_size(grouped_samples), ~ best_group_sizes(.x, n_min, n_max, n_ideal, prefer_big_groups))
-  names(subgroup_sizes) <- dplyr::group_keys(grouped_samples) %>%
-    tidyr::unite(col = "keys", sep = "/") %>%
+  names(subgroup_sizes) <- dplyr::group_keys(grouped_samples) |>
+    tidyr::unite(col = "keys", sep = "/") |>
     dplyr::pull(1)
 
   assertthat::assert_that(!strict || (min(unlist(subgroup_sizes)) >= n_min && max(unlist(subgroup_sizes)) <= n_max),
@@ -294,7 +294,7 @@ shuffle_with_subgroup_formation <- function(subgroup_object, subgroup_allocation
 
   # Index vector to assign group allocation to sample dataframe in its original order.
   # Split up according to group structure to allow easy permutation on the subgroup level
-  sample_vec <- order(dplyr::group_indices(subgroup_object$Grouped_Samples)) %>%
+  sample_vec <- order(dplyr::group_indices(subgroup_object$Grouped_Samples)) |>
     split(f = group_vec)
 
   # Order of the allocation variable in the batch container may be arbitrary and not strictly sorted;
@@ -330,7 +330,7 @@ shuffle_with_subgroup_formation <- function(subgroup_object, subgroup_allocation
 
     # Create a random permutation within (!) all the groups
     # This means that only 'equivalent' items are swapped around and get assigned to new subgroups every time
-    rand_index <- purrr::map(sample_vec, my_sample) %>% unlist(use.names = F)
+    rand_index <- purrr::map(sample_vec, my_sample) |> unlist(use.names = F)
 
     # Check if keep_separate_vars constraints are fulfilled within the randomized subgroups
     # Otherwise, create new permutations and increment number of tolerated violations every 1000 unsuccessful attempts
@@ -350,7 +350,7 @@ shuffle_with_subgroup_formation <- function(subgroup_object, subgroup_allocation
             "Increasing number of tolerated violations to ", sep_fails_tolerance
           )
         }
-        rand_index <- purrr::map(sample_vec, my_sample) %>% unlist(use.names = F)
+        rand_index <- purrr::map(sample_vec, my_sample) |> unlist(use.names = F)
       }
     }