+++ tests/testthat/test_optimize_design_simple_shuffle.R
@@ -6,11 +6,13 @@ samples <- data.frame(
   sampleId = seq_len(10)
 )
 
+env <- new.env()
+
 n_elements_changed <- function(bc) {
   df <- bc$get_samples(include_id = TRUE, as_tibble = FALSE)
   cur_state <- df$.sample_id
   cur_state <- tidyr::replace_na(cur_state, -1)
-  n_changed <<- c(n_changed, sum(start_state != cur_state))
+  env$n_changed <- c(env$n_changed, sum(start_state != cur_state))
   # never accept the change
   Inf
 }
@@ -23,8 +25,8 @@ set.seed(42)
 
 start_state <- ifelse(is.na(bc$assignment), -1, bc$assignment)
 
-n_changed <- numeric(0)
 test_that("correct number of shuffles = 1", {
+  env$n_changed <- numeric(0)
   optimize_design(
     bc,
     scoring = scoring_f,
@@ -32,11 +34,11 @@ test_that("correct number of shuffles = 1", {
     check_score_variance = F,
     autoscale_scores = F
   )
-  expect_equal(n_changed, c(0, rep(2, 10)))
+  expect_equal(env$n_changed, c(0, rep(2, 10)))
 })
 
-n_changed <- numeric(0)
 test_that("correct number of shuffles = 2", {
+  env$n_changed <- numeric(0)
   optimize_design(
     bc,
     scoring = scoring_f,
@@ -45,11 +47,11 @@ test_that("correct number of shuffles = 2", {
     check_score_variance = F,
     autoscale_scores = F
   )
-  expect_equal(n_changed, c(0, rep(4, 10)))
+  expect_equal(env$n_changed, c(0, rep(4, 10)))
 })
 
-n_changed <- numeric(0)
 test_that("correct number of shuffles = 5", {
+  env$n_changed <- numeric(0)
   optimize_design(
     bc,
     scoring = scoring_f,
@@ -58,11 +60,11 @@ test_that("correct number of shuffles = 5", {
     check_score_variance = FALSE,
     autoscale_scores = FALSE
   )
-  expect_equal(n_changed, c(0, rep(10, 10)))
+  expect_equal(env$n_changed, c(0, rep(10, 10)))
 })
 
-n_changed <- numeric(0)
 test_that("specify too many shuffles", {
+  env$n_changed <- numeric(0)
   optimize_design(
     bc,
     scoring = scoring_f,
@@ -71,11 +73,11 @@ test_that("specify too many shuffles", {
     check_score_variance = FALSE,
     autoscale_scores = FALSE
   )
-  expect_equal(n_changed, c(0, rep(20, 10)))
+  expect_equal(env$n_changed, c(0, rep(20, 10)))
 })
 
-n_changed <- numeric(0)
 test_that("complex shuffling schedule", {
+  env$n_changed <- numeric(0)
   optimize_design(
     bc,
     scoring = scoring_f,
@@ -84,5 +86,5 @@ test_that("complex shuffling schedule", {
     check_score_variance = F,
     autoscale_scores = F
   )
-  expect_equal(n_changed, c(0, c(4, 4, 10, 4, 4, 20, 20, 20, 20)))
+  expect_equal(env$n_changed, c(0, c(4, 4, 10, 4, 4, 20, 20, 20, 20)))
 })